% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_statistics.R
\name{compute_statistics}
\alias{compute_statistics}
\title{Summarise the iris dataset}
\usage{
compute_statistics(data, species)
}
\arguments{
\item{data}{a data frame}

\item{species}{a string}
}
\description{
Summarise the iris dataset
}
