% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_plot}
\alias{make_plot}
\title{Plot the correlation between iris flowers properties for one species}
\usage{
make_plot(statistics, species)
}
\arguments{
\item{statistics}{a data frame}

\item{species}{a string}
}
\description{
Plot the correlation between iris flowers properties for one species
}
