% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{create_plan}
\alias{create_plan}
\title{Create plan for Drake}
\usage{
create_plan(species = c("setosa", "versicolor", "virginica"), ...)
}
\arguments{
\item{species}{list of iris species}

\item{...}{other arguments for \link[drake:drake_plan]{drake_plan}}
}
\description{
Create plan for Drake
}
